Split up metric files.
